99_legacy_code/requirements.txt
fastapi
uvicorn[standard]

streamlit
st_clickable_images
streamlit_cropper
pandas
torch
transformers
faiss-cpu
